Restructure, all tests pass
